zoom in the map more, per #48
Here we're overriding the google.maps.Map.fitBounds() method and
shrinking the bounding box so that it zooms into a smaller region. The
math is a little wonky here because latitude isn't linear in Mercator,
but it appears to work well for all of the provinces and regions that
I've tested.