fix close bug on direct link to #about
If you go directly to #about in a new window/tab and close the dialog,
there won't be any history in the window's context, which redirects to
an empty page in Chrome. So now we check to see if there is history, and
if so we go(-1); otherwise we set location.hash to "" (or
options.closeHash).